PR review: notifications service (take2)
- simplify the pieces quite a bit

NOTE: This may not compile or run. I had no access to get GemFire config
required classes. The ideas are sound though - just may need to tweak the
syntax a bit.
